+++ src/snowflake/snowpark/dataframe.py
@@ -147,7 +147,6 @@ from snowflake.snowpark._internal.utils import (
     check_flatten_mode,
     column_to_bool,
     create_or_update_statement_params_with_query_tag,
-    create_prompt_column_from_template,
     deprecated,
     escape_quotes,
     experimental,
@@ -171,6 +170,7 @@ from snowflake.snowpark._internal.utils import (
 )
 from snowflake.snowpark.async_job import AsyncJob, _AsyncResultType
 from snowflake.snowpark.column import Column, _to_col_if_sql_expr, _to_col_if_str
+from snowflake.snowpark.dataframe_ai_functions import DataFrameAIFunctions
 from snowflake.snowpark.dataframe_analytics_functions import DataFrameAnalyticsFunctions
 from snowflake.snowpark.dataframe_na_functions import DataFrameNaFunctions
 from snowflake.snowpark.dataframe_stat_functions import DataFrameStatFunctions
@@ -191,7 +191,6 @@ from snowflake.snowpark.functions import (
     sql_expr,
     stddev,
     to_char,
-    ai_complete as ai_complete_function,
 )
 from snowflake.snowpark.mock._select_statement import MockSelectStatement
 from snowflake.snowpark.row import Row
@@ -650,6 +649,8 @@ class DataFrame:
         self.dropna = self._na.drop
         self.fillna = self._na.fill
         self.replace = self._na.replace
+
+        self._ai = DataFrameAIFunctions(self)
 
         self._alias: Optional[str] = None
 
@@ -4347,118 +4348,6 @@ class DataFrame:
 
         return df
 
-    @publicapi
-    @experimental(version="1.37.0")
-    def ai_complete(
-        self,
-        prompt: str,
-        input_columns: Union[List[Column], Dict[str, Column]],
-        output_column: Optional[str] = None,
-        *,
-        model: Optional[str] = None,
-        model_parameters: Optional[Dict[str, Any]] = None,
-        _emit_ast: bool = True,
-    ) -> "DataFrame":
-        """Generate a response (completion) on each row using the specified language model.
-
-        Args:
-            prompt: The prompt template string. Use placeholders like ``{name}`` when passing a dict of columns,
-                or ``{0}``, ``{1}`` when passing a list.
-            input_columns: A list of Columns (positional placeholders ``{0}``, ``{1}``, ...)
-                or a dict mapping placeholder names to Columns.
-            output_column: The name of the output column to be appended.
-                If not provided, a column named ``AI_COMPLETE_OUTPUT`` is appended.
-            model: Model name to pass to the underlying function.
-                It must be specified.
-            model_parameters: Optional dict containing model hyperparameters:
-
-                - temperature: Value from 0 to 1 controlling randomness (default: 0)
-                - top_p: Value from 0 to 1 controlling diversity (default: 0)
-                - max_tokens: Maximum number of output tokens (default: 4096, max: 8192)
-                - guardrails: Enable Cortex Guard filtering (default: False)
-
-        Returns:
-            A new DataFrame with appended output columns at the end.
-
-        Examples::
-
-            >>> # Single column output with named placeholder
-            >>> from snowflake.snowpark.functions import col
-            >>> df = session.create_dataframe(
-            ...     [["What is machine learning?"], ["Explain quantum computing"]],
-            ...     schema=["question"]
-            ... )
-            >>> result_df = df.ai_complete(
-            ...     prompt="Answer this question briefly: {q}",
-            ...     input_columns={"q": col("question")},
-            ...     output_column="answer",
-            ...     model="snowflake-arctic"
-            ... )
-            >>> result_df.columns
-            ['QUESTION', 'ANSWER']
-            >>> result_df.count()
-            2
-
-            >>> #  Processing images with file input
-            >>> from snowflake.snowpark.functions import to_file
-            >>> # Upload images to a stage first
-            >>> _ = session.sql("CREATE OR REPLACE TEMP STAGE mystage ENCRYPTION = (TYPE = 'SNOWFLAKE_SSE')").collect()
-            >>> _ = session.file.put("tests/resources/kitchen.png", "@mystage", auto_compress=False)
-            >>> _ = session.file.put("tests/resources/dog.jpg", "@mystage", auto_compress=False)
-            >>> # Create DataFrame with image paths and questions
-            >>> df = session.create_dataframe(
-            ...     [
-            ...         ["@mystage/kitchen.png", "What appliances are visible in this image?"],
-            ...         ["@mystage/dog.jpg", "What animal is in this image?"]
-            ...     ],
-            ...     schema=["image_path", "question"]
-            ... )
-            >>> # Use ai_complete with image files
-            >>> result_df = df.ai_complete(
-            ...     prompt="Image: {0}, Question: {1}",
-            ...     input_columns=[
-            ...         to_file(col("image_path")),
-            ...         col("question")
-            ...     ],
-            ...     output_column="answer",
-            ...     model="claude-4-sonnet"
-            ... )
-            >>> result_df.columns
-            ['IMAGE_PATH', 'QUESTION', 'ANSWER']
-            >>> result_df.count()
-            2
-            >>> results = result_df.collect()
-            >>> 'microwave' in results[0]["ANSWER"].lower()
-            True
-            >>> 'dog' in results[1]["ANSWER"].lower()
-            True
-        """
-
-        if not model:
-            raise ValueError("model must be specified for ai_complete")
-
-        # Build the prompt Column
-        if isinstance(input_columns, (dict, list)):
-            prompt_obj = create_prompt_column_from_template(
-                prompt, input_columns, _emit_ast=False
-            )
-        else:
-            raise TypeError(
-                "input_columns must be a list of Columns or a dict mapping placeholder names to Columns"
-            )
-
-        # Call the ai_complete function with all explicit parameters
-        result_col = ai_complete_function(
-            model=model,
-            prompt=prompt_obj,
-            model_parameters=model_parameters,
-            _emit_ast=False,
-        )
-
-        # Add the output column to the DataFrame
-        output_column_name = output_column or "AI_COMPLETE_OUTPUT"
-        return self.with_column(output_column_name, result_col, _emit_ast=False)
-
     @overload
     @publicapi
     def count(
@@ -5817,6 +5706,14 @@ class DataFrame:
         handling missing values in the DataFrame.
         """
         return self._na
+
+    @property
+    def ai(self) -> DataFrameAIFunctions:
+        """
+        Returns a :class:`DataFrameAIFunctions` object that provides AI-powered functions
+        for the DataFrame.
+        """
+        return self._ai
 
     @property
     def session(self) -> "snowflake.snowpark.Session":
